pyproject.toml
[project]
name = "yolo-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "onnx>=1.17.0",
    "onnxruntime>=1.21.0",
    "onnxslim>=0.1.49",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "ultralytics>=8.3.104",
]
